requirements.txt
langchain
langchain_community
langchain_google_genai
anthropic
python-dotenv
streamlit
youtube_transcript_api
yt-dlp
SpeechRecognition
pydub